+++ middlewares/ownership.js
@@ -2,7 +2,7 @@ const { verifyToken } = require('../modules/auth');
 const { roleAdmin } = require('../config/config').development;
 const { Forbidden } = require('../utils/status');
 const { throwError } = require('../utils/errorHandler');
-const { getTokenFromHeaders } = require('../utils/getTokenFromHeaders')
+const { getTokenFromHeaders } = require('../modules/auth');
 
 const ownershipValidator = async (req, res, next) => {
   const { id } = req.params;
