+++ modules/auth.js
@@ -1,5 +1,7 @@
 const jwt = require('jsonwebtoken');
 const { secretToken } = require('../config/config').development;
+const { Unauthorized } = require('./status');
+const { throwError } = require('./errorHandler');
 
 const createToken = (userInfo) => {
   const payload = {
@@ -18,4 +20,16 @@ const verifyToken = (token) => {
   return jwt.verify(token, secretToken);
 };
 
-module.exports = { createToken, verifyToken };
+const getTokenFromHeaders = (req) => {
+  const authorization = req.get('authorization') || req.headers['authorization'];
+
+  if (!authorization) {
+    throwError('No token provided', Unauthorized);
+  }
+
+  const tokenFormatting = authorization.substring(7);
+
+  return tokenFormatting;
+};
+
+module.exports = { createToken, verifyToken, getTokenFromHeaders };
